+++ demisto_sdk/commands/content_graph/parsers/json_content_item.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, List, Optional
 from demisto_sdk.commands.common.constants import (
     DEFAULT_CONTENT_ITEM_FROM_VERSION,
     DEFAULT_CONTENT_ITEM_TO_VERSION,
+    MINIMUM_XSOAR_SAAS_VERSION,
     MarketplaceVersions,
 )
 from demisto_sdk.commands.common.tools import get_json, get_value
@@ -75,7 +76,9 @@ class JSONContentItemParser(ContentItemParser):
         return get_value(
             self.json_data,
             self.field_mapping.get("fromversion", ""),
-            DEFAULT_CONTENT_ITEM_FROM_VERSION,
+            DEFAULT_CONTENT_ITEM_FROM_VERSION
+            if MarketplaceVersions.XSOAR_ON_PREM in self.supported_marketplaces
+            else MINIMUM_XSOAR_SAAS_VERSION,
         )
 
     @property